\name{rPDZ-package}
\alias{rPDZ-package}
\alias{rPDZ}
\docType{package}
\title{
  A utility for reading PDZ files into R
}
\description{
  This simple package allows X-ray fluorescence (XRF) spectra to be read into R
}
\details{
  There are two types of PDZ files, version 24 and 25. The first is straightforward to read in, the second requires flexible parsing of the binary.
}
\author{
Lee Drake, b.lee.drake@gmail.com

Maintainer: Lee Drake <b.lee.drake@gmail.com>
}
\references{
  https://www.xrf.guru
}
\keyword{ package }
\seealso{
  Optional links to other man pages
}
\examples{
  \dontrun{ 
  }
}
